+++ src/lib/models/book/ChapterID.ts
@@ -1,5 +1,21 @@
-const ChapterIDs = {
-    TableOfContentsID: '',
+export type ChapterIDKey =
+    | 'references'
+    | 'search'
+    | 'media'
+    | 'index'
+    | 'glossary'
+    | 'unknown'
+    | 'cover';
+
+export type TOCHeaderKey =
+    | 'chapters'
+    | 'acknowledgements'
+    | 'license'
+    | 'print'
+    | 'citation';
+
+const ChapterIDs: Record<string, ChapterIDKey> = {
+    TableOfContentsID: 'cover',
     ReferencesID: 'references',
     SearchID: 'search',
     MediaID: 'media',